# server/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for TaskPriority.
const (
	High   TaskPriority = "high"
	Low    TaskPriority = "low"
	Medium TaskPriority = "medium"
)

// Valid indicates whether the value is a known member of the TaskPriority enum.
func (e TaskPriority) Valid() bool {
	switch e {
	case High:
		return true
	case Low:
		return true
	case Medium:
		return true
	default:
		return false
	}
}

// Board defines model for Board.
type Board struct {
	Columns []Column `json:"columns"`
	Tasks   []Task   `json:"tasks"`
}

// Column defines model for Column.
type Column struct {
	Color string `json:"color"`
	Id    string `json:"id"`
	Name  string `json:"name"`
}

// CreateColumnRequest defines model for CreateColumnRequest.
type CreateColumnRequest struct {
	Name string `json:"name"`
}

// CreateTaskRequest defines model for CreateTaskRequest.
type CreateTaskRequest struct {
	ColumnId string `json:"columnId"`
}

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// MoveTaskRequest defines model for MoveTaskRequest.
type MoveTaskRequest struct {
	ColumnId string `json:"columnId"`
}

// Task defines model for Task.
type Task struct {
	Assignee string       `json:"assignee"`
	ColumnId string       `json:"columnId"`
	Id       int          `json:"id"`
	Key      string       `json:"key"`
	Label    string       `json:"label"`
	Priority TaskPriority `json:"priority"`
	Title    string       `json:"title"`
}

// TaskPriority defines model for Task.Priority.
type TaskPriority string

// UpdateColumnRequest defines model for UpdateColumnRequest.
type UpdateColumnRequest struct {
	Name string `json:"name"`
}

// BadRequest defines model for BadRequest.
type BadRequest = Error

// NotFound defines model for NotFound.
type NotFound = Error

// CreateColumnJSONRequestBody defines body for CreateColumn for application/json ContentType.
type CreateColumnJSONRequestBody = CreateColumnRequest

// RenameColumnJSONRequestBody defines body for RenameColumn for application/json ContentType.
type RenameColumnJSONRequestBody = UpdateColumnRequest

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = CreateTaskRequest

// MoveTaskJSONRequestBody defines body for MoveTask for application/json ContentType.
type MoveTaskJSONRequestBody = MoveTaskRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetBoard Get all board columns and tasks.
	// (GET /api/board)
	GetBoard(w http.ResponseWriter, r *http.Request)
	// CreateColumn Create a board column.
	// (POST /api/columns)
	CreateColumn(w http.ResponseWriter, r *http.Request)
	// DeleteColumn Delete a board column and move its tasks to another column.
	// (DELETE /api/columns/{columnId})
	DeleteColumn(w http.ResponseWriter, r *http.Request, columnId string)
	// RenameColumn Rename a board column.
	// (PATCH /api/columns/{columnId})
	RenameColumn(w http.ResponseWriter, r *http.Request, columnId string)
	// CreateTask Create a default task in a column.
	// (POST /api/tasks)
	CreateTask(w http.ResponseWriter, r *http.Request)
	// DeleteTask Delete a task.
	// (DELETE /api/tasks/{taskId})
	DeleteTask(w http.ResponseWriter, r *http.Request, taskId int)
	// MoveTask Move a task to another board column.
	// (PATCH /api/tasks/{taskId})
	MoveTask(w http.ResponseWriter, r *http.Request, taskId int)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// GetBoard Get all board columns and tasks.
// (GET /api/board)
func (_ Unimplemented) GetBoard(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreateColumn Create a board column.
// (POST /api/columns)
func (_ Unimplemented) CreateColumn(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// DeleteColumn Delete a board column and move its tasks to another column.
// (DELETE /api/columns/{columnId})
func (_ Unimplemented) DeleteColumn(w http.ResponseWriter, r *http.Request, columnId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// RenameColumn Rename a board column.
// (PATCH /api/columns/{columnId})
func (_ Unimplemented) RenameColumn(w http.ResponseWriter, r *http.Request, columnId string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// CreateTask Create a default task in a column.
// (POST /api/tasks)
func (_ Unimplemented) CreateTask(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// DeleteTask Delete a task.
// (DELETE /api/tasks/{taskId})
func (_ Unimplemented) DeleteTask(w http.ResponseWriter, r *http.Request, taskId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// MoveTask Move a task to another board column.
// (PATCH /api/tasks/{taskId})
func (_ Unimplemented) MoveTask(w http.ResponseWriter, r *http.Request, taskId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetBoard operation middleware
func (siw *ServerInterfaceWrapper) GetBoard(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBoard(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateColumn operation middleware
func (siw *ServerInterfaceWrapper) CreateColumn(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateColumn(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteColumn operation middleware
func (siw *ServerInterfaceWrapper) DeleteColumn(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "columnId" -------------
	var columnId string

	err = runtime.BindStyledParameterWithOptions("simple", "columnId", chi.URLParam(r, "columnId"), &columnId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "columnId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteColumn(w, r, columnId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RenameColumn operation middleware
func (siw *ServerInterfaceWrapper) RenameColumn(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "columnId" -------------
	var columnId string

	err = runtime.BindStyledParameterWithOptions("simple", "columnId", chi.URLParam(r, "columnId"), &columnId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "columnId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RenameColumn(w, r, columnId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateTask operation middleware
func (siw *ServerInterfaceWrapper) CreateTask(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateTask(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteTask operation middleware
func (siw *ServerInterfaceWrapper) DeleteTask(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "taskId" -------------
	var taskId int

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", chi.URLParam(r, "taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "taskId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteTask(w, r, taskId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MoveTask operation middleware
func (siw *ServerInterfaceWrapper) MoveTask(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "taskId" -------------
	var taskId int

	err = runtime.BindStyledParameterWithOptions("simple", "taskId", chi.URLParam(r, "taskId"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: r.URL.RawPath == ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "taskId", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MoveTask(w, r, taskId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/board", wrapper.GetBoard)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/columns", wrapper.CreateColumn)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/columns/{columnId}", wrapper.DeleteColumn)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/api/columns/{columnId}", wrapper.RenameColumn)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/tasks", wrapper.CreateTask)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/api/tasks/{taskId}", wrapper.DeleteTask)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/api/tasks/{taskId}", wrapper.MoveTask)
	})

	return r
}